refactor(workflow): split engine.go into focused files
1228-line engine.go → engine.go (orchestration core),
engine_steps.go (per-step-type executors), engine_retry.go
(backoff vars). No behavior changes; all tests pass.
